Move test to separate file. Need to think about codim-1 facet orientation..
